+++ packages/frontend/src/components/camera-check/CameraCheckActivity.component.tsx
@@ -1,8 +1,8 @@
 import { FC } from 'react';
 import { useNavigate } from 'react-router';
-import { ActivityCard } from './ActivityCard.component';
-import cameraCheckImage from '../assets/TaskPreviewImage.png';
-import { useTraining } from '../contexts/Training.contenxt';
+import { ActivityCard } from '../ActivityCard.component';
+import cameraCheckImage from '../../assets/TaskPreviewImage.png';
+import { useTraining } from '../../contexts/Training.contenxt';
 
 export const CameraCheckActivity: FC = () => {
   const navigate = useNavigate();
