+++ packages/frontend/src/components/practice/PracticeTaskInstructionsSide.component.tsx
