+++ packages/frontend/src/components/task/TaskInstructionsSide.component.tsx
@@ -1,19 +1,15 @@
+import { Stack, Typography, List, ListItem, ListItemText } from '@mui/material';
 import { FC } from 'react';
-import { Button, List, ListItem, ListItemText, Stack, Typography } from '@mui/material';
-import { TaskEntity } from '../client';
+import { TaskEntity } from '../../client';
 
-export interface TaskInstructionsProps {
+export interface TaskInstructionsSideProps {
   task: TaskEntity;
-  onStart: () => void;
 }
 
-export const TaskInstructions: FC<TaskInstructionsProps> = ({ task, onStart }) => {
+export const TaskInstructionsSide: FC<TaskInstructionsSideProps> = ({ task }) => {
   return (
-    <Stack sx={{ padding: 10 }}>
-      <Typography variant="h1">{task.title}</Typography>
-      <Typography variant="body1">{task.description}</Typography>
-
-      <Typography variant="h3" sx={{ paddingTop: 10 }}>The problem you have chosen is:</Typography>
+    <Stack>
+      <Typography variant="h3">The problem you have chosen is:</Typography>
       <Typography variant="body1">{task.problemDescription}</Typography>
 
       <Typography variant="h3" sx={{ paddingTop: 10 }}>Your task is to do the following:</Typography>
@@ -27,9 +23,6 @@ export const TaskInstructions: FC<TaskInstructionsProps> = ({ task, onStart }) =
           </ListItem>
         ))}
       </List>
-      <Button variant="contained" onClick={onStart}>
-        Start Recording
-      </Button>
     </Stack>
   );
 };
